C program that prints the first 50 Fibonacci numbers, starting with 1 and 2.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,14 +10,14 @@
 
 int main(void)
 {
-	int FN, FPO, FP1, i;
+	long int FN, FPO, FP1, i;
 
 	FPO = 1;
 	FP1 = 0;
 	for (i = 0; i < 50; i++)
 	{
 		FN = FP1 + FPO;
-		printf("%d, ", FN);
+		printf("%ld, ", FN);
 		FP1 = FPO;
 		FPO = FN;
 		FN = FP1 + FPO;
